# api/openapi.yaml
openapi: 3.1.0
info:
  title: BotHub API
  version: 0.1.0
  description: |
    Contract between the Go dashboard and the PHP API.

    Rules:
    - The API is the only way to change configuration. The dashboard never touches the DB.
    - Every route except `/api/health`, `GET /api/v1/setup`, `POST /api/v1/setup` and
      `POST /api/v1/auth/login` needs a valid session. One central middleware checks this.
    - Every changing request (POST, PUT, PATCH, DELETE) needs the header `X-CSRF-Token`.
    - Errors never contain UI text. They contain an i18n key that the dashboard translates.
    - Bot tokens are write-only. The API never returns them.
    - Actions that need Discord return `202` with a job. Poll `GET /api/v1/jobs/{jobId}`.
servers:
  - url: /
security:
  - session: []
...
paths:
...
  /api/v1/bots/{botId}/backups:
    parameters:
      - $ref: "#/components/parameters/BotId"
    get:
      operationId: listBotBackups
      tags: [bots]
      summary: Saved backups of the bot, global templates and ready-made templates (shared/bot-templates), without data
      responses:
        "200":
          description: Entries
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, kind, name, description, auto, createdAt, size]
                      properties:
                        id: { type: string, description: "Number, or builtin:<key> for a ready-made template" }
                        kind: { type: string, enum: [backup, template, builtin] }
                        name: { type: string }
                        description: { type: string }
                        auto: { type: boolean, description: Made automatically before a restore }
                        createdAt: { type: [string, "null"], format: date-time }
                        size: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: saveBotBackup
      tags: [bots]
      summary: Save the current state as a backup of this bot (max. 50) or as a global template (max. 100)
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                kind: { type: string, enum: [backup, template], default: backup }
                name: { type: string, minLength: 1, maxLength: 60 }
                description: { type: string, maxLength: 300 }
      responses:
        "201": { description: Saved }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/v1/bots/{botId}/backups/{backupId}:
    parameters:
      - $ref: "#/components/parameters/BotId"
      - { name: backupId, in: path, required: true, schema: { type: string, pattern: "^[a-z0-9:-]{1,60}$" } }
    get:
      operationId: getBotBackup
      tags: [bots]
      summary: Data of a saved entry or ready-made template (download)
      responses:
        "200": { description: Backup, content: { application/json: { schema: { type: object } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteBotBackup
      tags: [bots]
      summary: Delete a saved backup or template (ready-made templates cannot be deleted)
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204": { description: Deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/bots/{botId}/restore:
    parameters:
      - $ref: "#/components/parameters/BotId"
    post:
      operationId: restoreBot
      tags: [bots]
      summary: "Replaces the bot's configuration with a saved entry, a ready-made template or an uploaded backup; makes an automatic backup first. Every part is checked like in the editors (422 error.backup.part {part, reason})."
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string, description: Saved entry or builtin:<key> }
                data: { type: object, description: Uploaded backup file }
                name: { type: string, description: Name for the automatic backup }
      responses:
        "200": { description: Restored }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: bothub_session
...
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Token from `GET /api/v1/auth/me`
      schema: { type: string }
    BotId:
      name: botId
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
  responses:
    Unauthorized:
      description: No valid session (`error.auth.required`, `error.auth.invalid_credentials`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: CSRF token missing or wrong (`error.csrf.invalid`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found (`error.<resource>.not_found`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationError:
      description: Request body invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [key]
          properties:
            key:
              type: string
              description: i18n key, e.g. `error.bot.not_found`
              examples: [error.bot.not_found]
            params:
              type: object
              description: Values for placeholders in the translated text
              additionalProperties: true
            fields:
              type: object
              description: Per-field i18n keys for validation errors
              additionalProperties: { type: string }
...